+++ sflock/unpack/cab.py
@@ -1,4 +1,5 @@
 # Copyright (C) 2017-2018 Jurriaan Bremer.
+# Copyright (C) 2018 Hatching B.V.
 # This file is part of SFlock - http://www.sflock.org/.
 # See the file 'docs/LICENSE.txt' for copying permission.
 
@@ -24,14 +25,11 @@ class CabFile(Unpacker):
             filepath = self.f.temp_path(".cab")
             temporary = True
 
-        try:
-            subprocess.check_call([
-                self.zipjail, filepath, dirpath,
-                self.exe, "-d%s" % dirpath, filepath,
-            ], stdin=subprocess.PIPE, stdout=subprocess.PIPE)
-        except subprocess.CalledProcessError as e:
-            self.f.mode = "failed"
-            self.f.error = e
+        ret = self.zipjail(
+            filepath, dirpath, "-d%s" % dirpath, filepath
+        )
+        if not ret:
+            return []
 
         if temporary:
             os.unlink(filepath)
